Moving all the code to reasonable places
